requirements.txt
streamlit
pdfminer.six
python-dotenv
openai
tiktoken